+product functionality, location improvement
location now includes the name of the chain (at least as returned by
Kroger) and the street address.

Product functionality: Product search works. The program also accepts
product submissions from the client and is able to write them to disk
via the DBInterface.

Yet to send the stored product info to the client.